+++ examples/resources/harness_platform_gitops_project/resource.tf
@@ -1,7 +1,7 @@
 // Create gitOps project with account level agent
 resource "harness_platform_gitops_app_project" "test" {
   account_id = "account_id"
-  agent_id   = "agent_id"
+  agent_id   = "account.agent_id"
   upsert     = true
   project {
     metadata {